Permission's for SubDepartment,IncomeTax, Loan, LoanPaymentRecord, LoanType, MainAllowance, OverTimeType, PayrollName, PayrollPeriod, PayrollType, PayslipSetting, Position and TaxRegion

+++ app/Http/Controllers/Api/V1/OverTimeTypeController.php
@@ -7,35 +7,83 @@ use App\Models\Tenant\OverTimeType;
 use App\Http\Resources\Finance\OverTimeTypeResource;
 use App\Http\Requests\StoreOverTimeTypeRequest;
 use App\Http\Requests\UpdateOverTimeTypeRequest;
+use Illuminate\Http\Request;
+use Illuminate\Http\JsonResponse;
+use Spatie\Permission\Exceptions\PermissionDoesNotExist;
 
 class OverTimeTypeController extends Controller
 {
-    public function index()
+    public function index(Request $request)
     {
-        $overTimeTypes = OverTimeType::paginate(10);
-        return OverTimeTypeResource::collection($overTimeTypes);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('overtime_type_index')) {
+                $overTimeTypes = OverTimeType::paginate(10);
+                return OverTimeTypeResource::collection($overTimeTypes);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
     public function store(StoreOverTimeTypeRequest $request)
     {
-        $overTimeType = OverTimeType::create($request->validated());
-        return new OverTimeTypeResource($overTimeType);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('overtime_type_store')) {
+                $overTimeType = OverTimeType::create($request->validated());
+                return new OverTimeTypeResource($overTimeType);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
-    public function show(OverTimeType $overTimeType)
+    public function show(Request $request, OverTimeType $overTimeType)
     {
-        return new OverTimeTypeResource($overTimeType);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('overtime_type_show')) {
+                return new OverTimeTypeResource($overTimeType);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
     public function update(UpdateOverTimeTypeRequest $request, OverTimeType $overTimeType)
     {
-        $overTimeType->update($request->validated());
-        return new OverTimeTypeResource($overTimeType);
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('overtime_type_update')) {
+                $overTimeType->update($request->validated());
+                return new OverTimeTypeResource($overTimeType);
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 
-    public function destroy(OverTimeType $overTimeType)
+    public function destroy(Request $request, OverTimeType $overTimeType)
     {
-        $overTimeType->delete();
-        return response()->noContent();
+        try {
+            $user = $request->user();
+            if ($user->hasPermissionTo('overtime_type_destroy')) {
+                $overTimeType->delete();
+                return response()->noContent();
+            } else {
+                return response()->json(['message' => 'Unauthorized for this task'], 403);
+            }
+        } catch (PermissionDoesNotExist $exception) {
+            return response()->json(['message' => 'Unauthorized for this task - no permission by this name'], 403);
+        }
     }
 }
